Move the VulnerabilityAnalysisViewSet in the vulnerabilities module #103
