+++ dejacode/urls.py
@@ -51,8 +51,8 @@ from product_portfolio.api import ProductComponentViewSet
 from product_portfolio.api import ProductDependencyViewSet
 from product_portfolio.api import ProductPackageViewSet
 from product_portfolio.api import ProductViewSet
-from product_portfolio.api import VulnerabilityAnalysisViewSet
 from reporting.api import ReportViewSet
+from vulnerabilities.api import VulnerabilityAnalysisViewSet
 from vulnerabilities.api import VulnerabilityViewSet
 from workflow.api import RequestTemplateViewSet
 from workflow.api import RequestViewSet
